Перенес типы в отдельный файл
