+++ src/main/java/HTTPAPI.java
@@ -6,106 +6,9 @@ import java.net.URL;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
-import java.util.List;
 
 import org.apache.http.HttpStatus;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
-class GlvrdStatus implements GlvrdResponsable {
-    @JsonProperty("status")
-    public String status;
-
-    @JsonProperty("code")
-    public String code;
-
-    @JsonProperty("name")
-    public String name;
-
-    @JsonProperty("message")
-    public String message;
-
-    @JsonProperty("max_text_length")
-    public int max_text_length;
-
-    @JsonProperty("max_hints_count")
-    public int max_hints_count;
-
-    @JsonProperty("period_underlimit")
-    public boolean period_underlimit;
-
-    @JsonProperty("frequency_underlimit")
-    public boolean frequency_underlimit;
-
-    public String getStatus() {
-        return this.status;
-    }
-
-    public String getMessage() {
-        return this.message;
-    }
-}
-
-class GlvrdHTTPResponse implements GlvrdResponsable {
-    @JsonProperty("status")
-    public String status;
-
-    @JsonProperty("code")
-    public String code;
-
-    @JsonProperty("name")
-    public String name;
-
-    @JsonProperty("message")
-    public String message;
-
-    @JsonProperty("score")
-    public String score;
-
-    @JsonProperty("fragments")
-    public List<Fragment> fragments;
-
-    public String getStatus() {
-        return this.status;
-    }
-
-    public String getMessage() {
-        return this.message;
-    }
-}
-
-class Fragment {
-    @JsonProperty("start")
-    public int start;
-
-    @JsonProperty("end")
-    public int end;
-
-    @JsonProperty("hint_id")
-    public String hint_id;
-}
-
-class GlvrdHints implements GlvrdResponsable {
-    @JsonProperty("status")
-    public String status;
-
-    @JsonProperty("message")
-    public String message;
-
-    @JsonProperty("hints")
-    public JsonNode hints;
-
-    @Override
-    public String getStatus() {
-        return this.status;
-    }
-
-    @Override
-    public String getMessage() {
-        return this.message;
-    }
-}
 
 interface GlvrdResponsable {
     String getStatus();
